requirements.txt
psycopg2==2.9.9
mysql-connector-python==9.0.0
uuid==1.30
loguru==0.7.2